requirements.txt
djangorestframework>=3.12.0
djangorestframework-simplejwt>=4.7.2
django-cors-headers>=3.7.0
drf-spectacular
django-debug-toolbar==4.3.0